Hello .C
The cs50.h library is a custom library provided by the CS50 course which includes functions like get_string to simplify some tasks for the learners.

The get_string function is used to prompt the user for their name and store it in the string variable name. The %s format specifier in the printf function is used to display the value of the name variable as a string in the greeting message.


Finally, the return 0 statement is used to indicate that the program executed successfully.