Mario Less CS50x Project 1
This C code prompts the user to input a height for a pyramid and continues to re-prompt the user until a positive integer between 1 and 8 is entered. It then uses a loop to print a pyramid made of hashes (#) with each row having one more hash than the previous row, and is right-aligned with spaces. Finally, the program returns 0 to indicate successful completion.